+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 * TGR-745: Small performance boost for JEXL-execution
 * TGR-745: More compact Rbel-messages in WebUI
 * TGR-745: Ability added to filter TGR-files
+* TGR-758: Pairs are no longer discarded when filtering remote Traffic
 * TGR-766: Added support for Java 8 Date/Time Types to TigerGlobalConfiguration
 * TGR-575: Added support for CETP messages in tiger-rbel